+++ tests/tests_pytorch/trainer/connectors/test_accelerator_connector.py
@@ -41,7 +41,6 @@ from lightning.pytorch.accelerators.mps import MPSAccelerator
 from lightning.pytorch.plugins import DoublePrecisionPlugin, LayerSync, PrecisionPlugin, TorchSyncBatchNorm
 from lightning.pytorch.plugins.io import TorchCheckpointIO
 from lightning.pytorch.strategies import (
-    DDPSpawnStrategy,
     DDPStrategy,
     DeepSpeedStrategy,
     FSDPStrategy,
@@ -51,18 +50,13 @@ from lightning.pytorch.strategies import (
     SingleTPUStrategy,
     XLAStrategy,
 )
-from lightning.pytorch.strategies.ddp_spawn import _DDP_FORK_ALIASES
+from lightning.pytorch.strategies.ddp import _DDP_FORK_ALIASES
 from lightning.pytorch.strategies.hpu_parallel import HPUParallelStrategy
+from lightning.pytorch.strategies.launchers import _SubprocessScriptLauncher
 from lightning.pytorch.trainer.connectors.accelerator_connector import _set_torch_flags, AcceleratorConnector
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
 from tests_pytorch.conftest import mock_cuda_count, mock_mps_count, mock_tpu_available, mock_xla_available
 from tests_pytorch.helpers.runif import RunIf
-
-
-def test_accelerator_choice_cpu(tmpdir):
-    trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
-    assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, SingleDeviceStrategy)
 
 
 @RunIf(tpu=True, standalone=True)
@@ -96,12 +90,6 @@ def test_invalid_strategy_choice(invalid_strategy):
 def test_strategy_choice_ddp_on_cpu(tmpdir):
     """Test that selecting DDPStrategy on CPU works."""
     _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class=DDPStrategy)
-
-
-@RunIf(skip_windows=True)
-def test_strategy_choice_ddp_spawn_on_cpu(tmpdir):
-    """Test that selecting DDPSpawnStrategy on CPU works."""
-    _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class=DDPSpawnStrategy)
 
 
 def _test_strategy_choice_ddp_and_cpu(tmpdir, ddp_strategy_class):
@@ -235,14 +223,6 @@ def test_interactive_incompatible_backend_error(cuda_count_2, monkeypatch):
 
 
 @RunIf(skip_windows=True)
-@mock.patch("lightning.pytorch.strategies.xla.XLAStrategy.set_world_ranks")
-def test_interactive_compatible_strategy_tpu(_, tpu_available, monkeypatch):
-    monkeypatch.setattr(lightning.pytorch.trainer.connectors.accelerator_connector, "_IS_INTERACTIVE", True)
-    trainer = Trainer(accelerator="tpu")
-    assert trainer.strategy.launcher.is_interactive_compatible
-
-
-@RunIf(skip_windows=True)
 def test_interactive_compatible_strategy_ddp_fork(monkeypatch):
     monkeypatch.setattr(lightning.pytorch.trainer.connectors.accelerator_connector, "_IS_INTERACTIVE", True)
     trainer = Trainer(strategy="ddp_fork", accelerator="cpu")
@@ -254,7 +234,7 @@ def test_interactive_compatible_strategy_ddp_fork(monkeypatch):
     ["strategy", "strategy_class"],
     [
         ("ddp", DDPStrategy),
-        ("ddp_spawn", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
         pytest.param("deepspeed", DeepSpeedStrategy, marks=RunIf(deepspeed=True)),
     ],
 )
@@ -332,9 +312,9 @@ def test_set_devices_if_none_cpu():
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     (
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_true", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_true", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
         ("ddp_find_unused_parameters_true", DDPStrategy),
@@ -353,8 +333,8 @@ def test_invalid_ddp_strategy_with_mps(accelerator, strategy, strategy_class, mp
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     [
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
     ],
@@ -364,17 +344,16 @@ def test_strategy_choice_cpu_str(strategy, strategy_class):
     assert isinstance(trainer.strategy, strategy_class)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_strategy_choice_cpu_instance(strategy_class):
-    trainer = Trainer(strategy=strategy_class(), accelerator="cpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_strategy_choice_cpu_instance():
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="cpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
 
 
 @pytest.mark.parametrize(
     ["strategy", "strategy_class"],
     [
-        ("ddp_spawn", DDPSpawnStrategy),
-        ("ddp_spawn_find_unused_parameters_false", DDPSpawnStrategy),
+        ("ddp_spawn", DDPStrategy),
+        ("ddp_spawn_find_unused_parameters_false", DDPStrategy),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
         pytest.param("deepspeed", DeepSpeedStrategy, marks=RunIf(deepspeed=True)),
@@ -385,16 +364,14 @@ def test_strategy_choice_gpu_str(strategy, strategy_class, cuda_count_2, mps_cou
     assert isinstance(trainer.strategy, strategy_class)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_strategy_choice_gpu_instance(strategy_class, cuda_count_2, mps_count_0):
-    trainer = Trainer(strategy=strategy_class(), accelerator="gpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_strategy_choice_gpu_instance(cuda_count_2, mps_count_0):
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="gpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
 
 
-@pytest.mark.parametrize("strategy_class", [DDPSpawnStrategy, DDPStrategy])
-def test_device_type_when_strategy_instance_gpu_passed(strategy_class, cuda_count_2, mps_count_0):
-    trainer = Trainer(strategy=strategy_class(), accelerator="gpu", devices=2)
-    assert isinstance(trainer.strategy, strategy_class)
+def test_device_type_when_strategy_instance_gpu_passed(cuda_count_2, mps_count_0):
+    trainer = Trainer(strategy=DDPStrategy(), accelerator="gpu", devices=2)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.accelerator, CUDAAccelerator)
 
 
@@ -408,7 +385,7 @@ def test_validate_precision_type(precision):
 def test_strategy_choice_ddp_spawn_cpu():
     trainer = Trainer(strategy="ddp_spawn", accelerator="cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "spawn"
 
@@ -416,11 +393,11 @@ def test_strategy_choice_ddp_spawn_cpu():
 @RunIf(skip_windows=True)
 @mock.patch("lightning.pytorch.trainer.connectors.accelerator_connector._IS_INTERACTIVE", True)
 def test_strategy_choice_ddp_fork_in_interactive():
-    """Test that when accelerator and strategy are unspecified, the connector chooses DDP Fork in interactive
-    environments by default."""
-    trainer = Trainer(devices=2)
+    """Test that when strategy is unspecified, the connector chooses DDP Fork in interactive environments by
+    default."""
+    trainer = Trainer(accelerator="cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "fork"
 
@@ -429,12 +406,12 @@ def test_strategy_choice_ddp_fork_in_interactive():
 def test_strategy_choice_ddp_fork_cpu():
     trainer = Trainer(strategy="ddp_fork", accelerator="cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert isinstance(trainer.strategy, DDPSpawnStrategy)
+    assert isinstance(trainer.strategy, DDPStrategy)
     assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
     assert trainer.strategy.launcher._start_method == "fork"
 
 
-@pytest.mark.parametrize("strategy,expected_cls", [("ddp", DDPStrategy), ("ddp_spawn", DDPSpawnStrategy)])
+@pytest.mark.parametrize("strategy,expected_cls", [("ddp", DDPStrategy), ("ddp_spawn", DDPStrategy)])
 def test_strategy_choice_ddp_cuda(strategy, expected_cls, mps_count_0, cuda_count_2):
     trainer = Trainer(fast_dev_run=True, strategy=strategy, accelerator="gpu", devices=1)
     assert isinstance(trainer.accelerator, CUDAAccelerator)
@@ -601,20 +578,6 @@ def test_devices_auto_choice_cpu(monkeypatch, cuda_count_0):
     mock_hpu_available(monkeypatch, False)
     mock_xla_available(monkeypatch, False)
     trainer = Trainer(accelerator="auto", devices="auto")
-    assert trainer.num_devices == 1
-
-
-@RunIf(mps=False)
-def test_devices_auto_choice_gpu(cuda_count_2):
-    trainer = Trainer(accelerator="auto", devices="auto")
-    assert isinstance(trainer.accelerator, CUDAAccelerator)
-    assert trainer.num_devices == 2
-
-
-@RunIf(mps=True)
-def test_devices_auto_choice_mps():
-    trainer = Trainer(accelerator="auto", devices="auto")
-    assert isinstance(trainer.accelerator, MPSAccelerator)
     assert trainer.num_devices == 1
 
 
@@ -826,8 +789,19 @@ def test_colossalai_external_strategy(monkeypatch):
     assert isinstance(trainer.strategy, ColossalAIStrategy)
 
 
-def test_connector_auto_selection(monkeypatch):
+@pytest.mark.parametrize("is_interactive", (False, True))
+def test_connector_auto_selection(monkeypatch, is_interactive):
     import lightning.fabric  # avoid breakage with standalone package
+
+    def _mock_interactive():
+        monkeypatch.setattr(
+            lightning.pytorch.trainer.connectors.accelerator_connector, "_IS_INTERACTIVE", is_interactive
+        )
+        if _IS_WINDOWS:
+            # simulate fork support on windows
+            monkeypatch.setattr(torch.multiprocessing, "get_all_start_methods", lambda: ["fork", "spawn"])
+
+    _mock_interactive()
 
     def _mock_tpu_available(value):
         mock_tpu_available(monkeypatch, value)
@@ -842,10 +816,11 @@ def test_connector_auto_selection(monkeypatch):
         mock_tpu_available(monkeypatch, False)
         mock_ipu_available(monkeypatch, False)
         mock_hpu_available(monkeypatch, False)
-        connector = AcceleratorConnector()
-    assert isinstance(connector.accelerator, CPUAccelerator)
-    assert isinstance(connector.strategy, SingleDeviceStrategy)
-    assert connector._devices_flag == 1
+        trainer = Trainer()
+    assert isinstance(trainer.accelerator, CPUAccelerator)
+    assert isinstance(trainer.strategy, SingleDeviceStrategy)
+    assert trainer._accelerator_connector._devices_flag == 1
+    assert trainer.num_devices == 1
 
     # single CUDA
     with monkeypatch.context():
@@ -854,10 +829,11 @@ def test_connector_auto_selection(monkeypatch):
         mock_tpu_available(monkeypatch, False)
         mock_ipu_available(monkeypatch, False)
         mock_hpu_available(monkeypatch, False)
-        connector = AcceleratorConnector()
-    assert isinstance(connector.accelerator, CUDAAccelerator)
-    assert isinstance(connector.strategy, SingleDeviceStrategy)
-    assert connector._devices_flag == [0]
+        trainer = Trainer()
+    assert isinstance(trainer.accelerator, CUDAAccelerator)
+    assert isinstance(trainer.strategy, SingleDeviceStrategy)
+    assert trainer._accelerator_connector._devices_flag == [0]
+    assert trainer.num_devices == 1
 
     # multi CUDA
     with monkeypatch.context():
@@ -866,10 +842,14 @@ def test_connector_auto_selection(monkeypatch):
         mock_tpu_available(monkeypatch, False)
         mock_ipu_available(monkeypatch, False)
         mock_hpu_available(monkeypatch, False)
-        connector = AcceleratorConnector()
-    assert isinstance(connector.accelerator, CUDAAccelerator)
-    assert isinstance(connector.strategy, DDPStrategy)
-    assert connector._devices_flag == list(range(4))
+        trainer = Trainer()
+    assert isinstance(trainer.accelerator, CUDAAccelerator)
+    assert isinstance(trainer.strategy, DDPStrategy)
+    assert trainer._accelerator_connector._devices_flag == list(range(4))
+    assert trainer.num_devices == 4
+    assert isinstance(trainer.strategy.cluster_environment, LightningEnvironment)
+    assert trainer.strategy._start_method == ("fork" if is_interactive else "popen")
+    assert trainer.strategy.launcher.is_interactive_compatible == is_interactive
 
     # MPS (there's no distributed)
     with monkeypatch.context():
@@ -900,12 +880,10 @@ def test_connector_auto_selection(monkeypatch):
 
     monkeypatch.undo()  # for some reason `.context()` is not working properly
     assert lightning.fabric.accelerators.TPUAccelerator.auto_device_count() == 8
+    _mock_interactive()
 
     # Multi TPU
     with monkeypatch.context():
-        if _IS_WINDOWS:
-            # simulate fork support on windows
-            monkeypatch.setattr(torch.multiprocessing, "get_all_start_methods", lambda: ["fork", "spawn"])
         mock_cuda_count(monkeypatch, 0)
         mock_mps_count(monkeypatch, 0)
         _mock_tpu_available(True)
@@ -915,6 +893,9 @@ def test_connector_auto_selection(monkeypatch):
     assert isinstance(connector.accelerator, TPUAccelerator)
     assert isinstance(connector.strategy, XLAStrategy)
     assert connector._devices_flag == 8
+    assert isinstance(connector.strategy.cluster_environment, XLAEnvironment)
+    assert connector.strategy._start_method == "fork"
+    assert connector.strategy.launcher.is_interactive_compatible
 
     # Single/Multi IPU: strategy is the same
     with monkeypatch.context():
@@ -927,6 +908,8 @@ def test_connector_auto_selection(monkeypatch):
     assert isinstance(connector.accelerator, IPUAccelerator)
     assert isinstance(connector.strategy, IPUStrategy)
     assert connector._devices_flag == 4
+    assert isinstance(connector.strategy.cluster_environment, LightningEnvironment)
+    assert connector.strategy.launcher is None
 
     # Single HPU
     with monkeypatch.context():
@@ -942,24 +925,26 @@ def test_connector_auto_selection(monkeypatch):
     assert connector._devices_flag == 1
 
     monkeypatch.undo()  # for some reason `.context()` is not working properly
+    _mock_interactive()
 
-    # Multi HPU
-    with monkeypatch.context():
-        mock_cuda_count(monkeypatch, 0)
-        mock_mps_count(monkeypatch, 0)
-        mock_tpu_available(monkeypatch, False)
-        mock_ipu_available(monkeypatch, False)
-        mock_hpu_available(monkeypatch, True)
-        connector = AcceleratorConnector()
-    assert isinstance(connector.accelerator, HPUAccelerator)
-    assert isinstance(connector.strategy, HPUParallelStrategy)
-    assert connector._devices_flag == 8
+    if not is_interactive:  # HPU does not support interactive environments
+        # Multi HPU
+        with monkeypatch.context():
+            mock_cuda_count(monkeypatch, 0)
+            mock_mps_count(monkeypatch, 0)
+            mock_tpu_available(monkeypatch, False)
+            mock_ipu_available(monkeypatch, False)
+            mock_hpu_available(monkeypatch, True)
+            connector = AcceleratorConnector()
+        assert isinstance(connector.accelerator, HPUAccelerator)
+        assert isinstance(connector.strategy, HPUParallelStrategy)
+        assert connector._devices_flag == 8
+        assert isinstance(connector.strategy.cluster_environment, LightningEnvironment)
+        assert isinstance(connector.strategy.launcher, _SubprocessScriptLauncher)
+        assert not connector.strategy.launcher.is_interactive_compatible
 
     # TPU and CUDA: prefers TPU
     with monkeypatch.context():
-        if _IS_WINDOWS:
-            # simulate fork support on windows
-            monkeypatch.setattr(torch.multiprocessing, "get_all_start_methods", lambda: ["fork", "spawn"])
         mock_cuda_count(monkeypatch, 2)
         mock_mps_count(monkeypatch, 0)
         _mock_tpu_available(True)
@@ -969,3 +954,6 @@ def test_connector_auto_selection(monkeypatch):
     assert isinstance(connector.accelerator, TPUAccelerator)
     assert isinstance(connector.strategy, XLAStrategy)
     assert connector._devices_flag == 8
+    assert isinstance(connector.strategy.cluster_environment, XLAEnvironment)
+    assert connector.strategy._start_method == "fork"
+    assert connector.strategy.launcher.is_interactive_compatible